Split Arena and Player classes to different files
